Backend.Tests.End2End: add lnd->geewallet test
Add a test which opens a channel from lnd to geewallet.

We also raise RegTest fee estimation, otherwise the AcceptChannel test
would fail with:
1) Test Error : GWallet.Backend.Tests.LN.can accept channel from LND
   System.Threading.Tasks.TaskCanceledException : A task was canceled.

(Note that we also add the Ignore attrib cause still after raising
the fee, the test doesn't pass. It will be fixed in the subsequent
commits.)