End2End: lnd.FundByMining mine to throw-away address
Blocks mined to geewallet's address in end-to-end tests aren't
recognised as spendable txouts. When we were mining to geewallet's
address we were only using it as a dummy address because we didn't care
where the funds went. This caused problems however: it made the
intention of the code unclear (by making the use of geewallet's address
seem relevant), it would also break tests if you try to replace
electrumx with electrs (by making electrs fail to report the wallet
balance).

We now mine to a throw-away address instead of using geewallet's
address.