Backend,CI: new e2e pipelines (with RegTest deps)
The currentDir change in make.fsx is done so that the text
search is not done in any of the files installed by these deps.

The Backend fixes are three (which should have been done in the
commit that added RegTest support[1] but the need for them was
only exposed by failures in this end2end test).

1) baseline fix: otherwise we would get this exception:

"Merge&Save didn't happen before launching the
FaultTolerantPClient?"

2) averageFunc fix: otherwise we would get this exception:

"numberOfParallelJobsAllowed should be equal or higher than
minimumNumberOfResponses for the averageFunc"

3) fee estimation fix: otherwise we would get this exception:

"GWallet.Backend.ServerMisconfiguredException: Fee estimation
returned a -1 error code"

[1] b6bd065650e0e8a05d758704ba64e0e7c4e4f8fd